+++ packages/webblocks-cms/src/Http/Controllers/Admin/PageController.php
@@ -1029,7 +1029,7 @@ class PageController extends Controller
 
     if (! $parentId) {
       return $blockTypes
-        ->reject(fn (BlockType $blockType) => in_array($blockType->slug, ['sidebar-nav-item', 'sidebar-nav-group', 'card_header', 'card_body', 'card_footer'], true))
+        ->reject(fn (BlockType $blockType) => in_array($blockType->slug, ['sidebar-nav-item', 'sidebar-nav-group', 'card_header', 'card_body', 'card_footer', 'slide'], true))
         ->values();
     }